fix: Validation of git models
