+++ backend/capellacollab/projects/toolmodels/diagrams/validation.py
@@ -6,15 +6,15 @@ import logging
 from sqlalchemy import orm
 
 import capellacollab.projects.toolmodels.models as toolmodels_models
-import capellacollab.projects.toolmodels.modelsources.git.gitlab.models as gitlab_models
-import capellacollab.projects.toolmodels.modelsources.git.gitlab.validation as gitlab_validation
+import capellacollab.projects.toolmodels.modelsources.git.models as git_models
+import capellacollab.projects.toolmodels.modelsources.git.validation as git_validation
 
 
-def check_diagram_cache_health(
+async def check_diagram_cache_health(
     db: orm.Session,
     model: toolmodels_models.DatabaseCapellaModel,
     logger: logging.LoggerAdapter,
-) -> gitlab_models.ModelArtifactStatus:
-    return gitlab_validation.check_pipeline_health(
+) -> git_models.ModelArtifactStatus:
+    return await git_validation.check_pipeline_health(
         db, model, "update_capella_diagram_cache", logger
     )
