+++ backend/capellacollab/projects/toolmodels/modelsources/git/github/handler.py
@@ -95,7 +95,7 @@ class GithubHandler(handler.GitHandler):
     ) -> t.Any:
         headers = None
         if not self.git_model.password:
-            headers = self.get_headers(self.git_model.password)
+            headers = self.__get_headers(self.git_model.password)
 
         response = requests.get(
             f"{self.git_instance.api_url}/repos{project_id}/actions/runs?branch={parse.quote(self.git_model.revision, safe='')}&per_page=20",
